Django-PlayGround/Django Fun/requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
black==22.3.0
click==8.1.2
Deprecated==1.2.13
Django==4.0.3
django-filter==21.1
django-redis==5.2.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
Faker==13.3.4
Markdown==3.3.6
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.2.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
validate-cpf==0.1.2
validate-docbr==1.10.0
wrapt==1.14.0
